Changed Twitter service to only provide English results and refactor hometeam tweets job
