+++ src/main/java/com/codingnomads/betty/data/batch/tweetsjobs/hometeamjob/HomeTeamTweetsRdsWriter.java
@@ -1,4 +1,4 @@
-package com.codingnomads.betty.data.batch.tweetsjob.hometeamstep;
+package com.codingnomads.betty.data.batch.tweetsjobs.hometeamjob;
 
 import com.codingnomads.betty.data.models.Tweet;
 import com.codingnomads.betty.logic.interfaces.TwitterJpaRepository;
